pyproject.toml
[project]
description = "An API for admin of sporting users"
license = "Apache-2.0"
name = "user-service"
readme = "README.md"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiodns>=3.2.0",
    "aiohttp-middlewares>=2.4.0",
    "aiohttp>=3.10.10",
    "certifi>=2024.8.30",
    "dataclasses-json>=0.6.7",
    "gunicorn>=23.0.0",
    "motor>=3.6.0",
    "multidict>=6.1.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "aiohttp-devtools>=1.1.2",
    "poethepoet>=0.29.0",
    "pyright>=1.1.386",
    "pytest-aiohttp>=1.0.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-docker>=3.1.1",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
    "pytest>=8.3.3",
    "requests>=2.32.3",
    "ruff>=0.7.1",
]

[tool.poe.tasks]
fmt = "uv run ruff format"
lint = "uv run ruff check --fix"
pyright = "uv run pyright"
unit-tests = "uv run pytest -m unit"
integration-tests = "uv run pytest --cov=user_service --cov-report=term-missing -m integration"
contract-tests = "uv run pytest -m contract"
release = [
    "lint",
    "pyright",
    "unit-tests",
    "integration-tests",
    "contract-tests",
]


[tool.ruff.lint]
select = ["ALL"]
# and then manually ignore annoying ones:
ignore = [
    "ANN401", # Any should be allowed
    "D211",
    "D213",
    "COM812",
    "D203",
    "ISC001",
]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    # at least this three should be fine in tests:
    "S101", # asserts allowed in tests...
    #     "ARG",  # Unused function args -> fixtures nevertheless are functionally relevant...
    #     "FBT",  # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
]

[tool.ruff.lint.isort]
# so it knows to group first-party stuff last
known-first-party = ["user_service"]

[tool.pyright]
venvPath = "." # rye installs the venv in the current dir
venv = ".venv" #       in a folder called `.venv`
#strict = ["**/*.py"] # use 'strict' checking on all files

[tool.pytest.ini_options]
env = [
    "CONFIG=test",
    "JWT_SECRET=secret",
    "ADMIN_USERNAME=admin",
    "ADMIN_PASSWORD=password",
    "JWT_EXP_DELTA_SECONDS=60",
    "JWT_SECRET=secret",
    "DB_NAME=users_test",
    "DB_USER=admin",
    "DB_PASSWORD=admin",
    "LOGGING_LEVEL=DEBUG",
]
asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit",
    "integration: marks tests as integration",
    "contract: marks tests as contract",
]

[tool.coverage.paths]
source = ["user_service", "*/site-packages"]

[tool.coverage.run]
branch = false
omit = ["*gunicorn_config.py", "users_adapter.py", "adapter.py"]
source = ["user_service"]

[tool.coverage.report]
fail_under = 100
show_missing = true
